+++ streams/sqlite/voxel_stream_sqlite.h
@@ -23,11 +23,11 @@ public:
 	void set_database_path(String path);
 	String get_database_path() const;
 
-	Result emerge_block(VoxelBufferInternal &out_buffer, Vector3i origin_in_voxels, int lod) override;
-	void immerge_block(VoxelBufferInternal &buffer, Vector3i origin_in_voxels, int lod) override;
+	Result load_voxel_block(VoxelBufferInternal &out_buffer, Vector3i origin_in_voxels, int lod) override;
+	void save_voxel_block(VoxelBufferInternal &buffer, Vector3i origin_in_voxels, int lod) override;
 
-	void emerge_blocks(Span<VoxelBlockRequest> p_blocks, Vector<Result> &out_results) override;
-	void immerge_blocks(Span<VoxelBlockRequest> p_blocks) override;
+	void load_voxel_blocks(Span<VoxelBlockRequest> p_blocks, Vector<Result> &out_results) override;
+	void save_voxel_blocks(Span<VoxelBlockRequest> p_blocks) override;
 
 	bool supports_instance_blocks() const override;
 	void load_instance_blocks(Span<VoxelStreamInstanceDataRequest> out_blocks, Span<Result> out_results) override;