+++ streams/voxel_stream_script.cpp
@@ -4,7 +4,7 @@
 
 namespace zylann::voxel {
 
-VoxelStream::Result VoxelStreamScript::emerge_block(
+VoxelStream::Result VoxelStreamScript::load_voxel_block(
 		VoxelBufferInternal &out_buffer, Vector3i origin_in_voxels, int lod) {
 	Variant output;
 	// Create a temporary wrapper so Godot can pass it to scripts
@@ -15,7 +15,7 @@ VoxelStream::Result VoxelStreamScript::emerge_block(
 
 	//Result res;
 	int res;
-	if (GDVIRTUAL_CALL(_emerge_block, buffer_wrapper, origin_in_voxels, lod, res)) {
+	if (GDVIRTUAL_CALL(_load_voxel_block, buffer_wrapper, origin_in_voxels, lod, res)) {
 		ERR_FAIL_INDEX_V(res, _RESULT_COUNT, RESULT_ERROR);
 		return Result(res);
 	} else {
@@ -27,11 +27,11 @@ VoxelStream::Result VoxelStreamScript::emerge_block(
 	return RESULT_ERROR;
 }
 
-void VoxelStreamScript::immerge_block(VoxelBufferInternal &buffer, Vector3i origin_in_voxels, int lod) {
+void VoxelStreamScript::save_voxel_block(VoxelBufferInternal &buffer, Vector3i origin_in_voxels, int lod) {
 	Ref<VoxelBuffer> buffer_wrapper;
 	buffer_wrapper.instantiate();
 	buffer.duplicate_to(buffer_wrapper->get_buffer(), true);
-	if (!GDVIRTUAL_CALL(_immerge_block, buffer_wrapper, origin_in_voxels, lod)) {
+	if (!GDVIRTUAL_CALL(_save_voxel_block, buffer_wrapper, origin_in_voxels, lod)) {
 		WARN_PRINT_ONCE("VoxelStreamScript::_immerge_block is unimplemented!");
 	}
 }
@@ -46,8 +46,8 @@ int VoxelStreamScript::get_used_channels_mask() const {
 
 void VoxelStreamScript::_bind_methods() {
 	// TODO Test if GDVIRTUAL can print errors properly when GDScript fails inside a different thread.
-	GDVIRTUAL_BIND(_emerge_block, "out_buffer", "origin_in_voxels", "lod");
-	GDVIRTUAL_BIND(_immerge_block, "buffer", "origin_in_voxels", "lod");
+	GDVIRTUAL_BIND(_load_voxel_block, "out_buffer", "origin_in_voxels", "lod");
+	GDVIRTUAL_BIND(_save_voxel_block, "buffer", "origin_in_voxels", "lod");
 	GDVIRTUAL_BIND(_get_used_channels_mask);
 
 	// BIND_VMETHOD(MethodInfo("_emerge_block",