+++ streams/voxel_stream_script.h
@@ -13,16 +13,16 @@ namespace zylann::voxel {
 class VoxelStreamScript : public VoxelStream {
 	GDCLASS(VoxelStreamScript, VoxelStream)
 public:
-	Result emerge_block(VoxelBufferInternal &out_buffer, Vector3i origin_in_voxels, int lod) override;
-	void immerge_block(VoxelBufferInternal &buffer, Vector3i origin_in_voxels, int lod) override;
+	Result load_voxel_block(VoxelBufferInternal &out_buffer, Vector3i origin_in_voxels, int lod) override;
+	void save_voxel_block(VoxelBufferInternal &buffer, Vector3i origin_in_voxels, int lod) override;
 
 	int get_used_channels_mask() const override;
 
 protected:
 	// TODO Why is it unable to convert `Result` into `Variant` even though a cast is defined in voxel_stream.h???
 	//GDVIRTUAL3R(VoxelStream::Result, _emerge_block, Ref<VoxelBuffer>, Vector3i, int)
-	GDVIRTUAL3R(int, _emerge_block, Ref<VoxelBuffer>, Vector3i, int)
-	GDVIRTUAL3(_immerge_block, Ref<VoxelBuffer>, Vector3i, int)
+	GDVIRTUAL3R(int, _load_voxel_block, Ref<VoxelBuffer>, Vector3i, int)
+	GDVIRTUAL3(_save_voxel_block, Ref<VoxelBuffer>, Vector3i, int)
 	GDVIRTUAL0RC(int, _get_used_channels_mask) // I think `C` means `const`?
 
 	static void _bind_methods();